fixed imports for moved Apis
