+++ fbw-a32nx/src/systems/instruments/src/EFB/Navigation/Navigation.tsx
@@ -17,7 +17,7 @@ import { ReactZoomPanPinchRef, TransformComponent, TransformWrapper } from 'reac
 import { t } from '../translation';
 import { TooltipWrapper } from '../UtilComponents/TooltipWrapper';
 // import { DrawableCanvas } from '../UtilComponents/DrawableCanvas';
-import { useNavigraph } from '../ChartsApi/Navigraph';
+import { useNavigraph } from '../Apis/Navigraph/Navigraph';
 import { SimpleInput } from '../UtilComponents/Form/SimpleInput/SimpleInput';
 import { useAppDispatch, useAppSelector } from '../Store/store';
 import {
@@ -210,7 +210,7 @@ export const ChartViewer = () => {
             >
                 {isFullScreen && (
                     <div
-                        className="flex absolute top-6 right-6 flex-row items-center p-4 hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight rounded-md transition duration-100"
+                        className="flex absolute top-6 right-6 flex-row items-center p-4 rounded-md transition duration-100 hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight"
                         onClick={() => dispatch(editTabProperty({ tab: currentTab, isFullScreen: false }))}
                     >
                         <FullscreenExit size={40} />
@@ -342,7 +342,7 @@ export const ChartViewer = () => {
                                                 }));
                                             }
                                         }}
-                                        className="p-2 hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight transition duration-100 cursor-pointer"
+                                        className="p-2 transition duration-100 cursor-pointer hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight"
                                     >
                                         <ArrowsExpand size={40} />
                                     </button>
@@ -389,7 +389,7 @@ export const ChartViewer = () => {
                                                 }));
                                             }
                                         }}
-                                        className="p-2 hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight transition duration-100 cursor-pointer"
+                                        className="p-2 transition duration-100 cursor-pointer hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight"
                                     >
                                         <ArrowsExpand className="transform rotate-90" size={40} />
                                     </button>
@@ -403,7 +403,7 @@ export const ChartViewer = () => {
                                             dispatch(editTabProperty({ tab: currentTab, chartRotation: 0 }));
                                             dispatch(editTabProperty({ tab: currentTab, chartPosition: { ...chartPosition, positionX: 0, positionY: 0, scale: 1 } }));
                                         }}
-                                        className="p-2 hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight transition duration-100 cursor-pointer"
+                                        className="p-2 transition duration-100 cursor-pointer hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight"
                                     >
                                         <XCircleFill size={40} />
                                     </button>
@@ -413,7 +413,7 @@ export const ChartViewer = () => {
                                     <button
                                         type="button"
                                         onClick={() => zoomIn()}
-                                        className="p-2 hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight transition duration-100 cursor-pointer"
+                                        className="p-2 transition duration-100 cursor-pointer hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight"
                                     >
                                         <Plus size={40} />
                                     </button>
@@ -423,7 +423,7 @@ export const ChartViewer = () => {
                                     <button
                                         type="button"
                                         onClick={() => zoomOut()}
-                                        className="p-2 hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight transition duration-100 cursor-pointer"
+                                        className="p-2 transition duration-100 cursor-pointer hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight"
                                     >
                                         <Dash size={40} />
                                     </button>
@@ -431,7 +431,7 @@ export const ChartViewer = () => {
                             </div>
                             <div className="flex overflow-hidden flex-col rounded-md">
                                 <div
-                                    className="p-2 hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight rounded-md transition duration-100 cursor-pointer"
+                                    className="p-2 rounded-md transition duration-100 cursor-pointer hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight"
                                     onClick={() => {
                                         dispatch(editTabProperty({ tab: currentTab, isFullScreen: !isFullScreen }));
                                         if (chartRef.current && ref.current) {
@@ -447,7 +447,7 @@ export const ChartViewer = () => {
 
                                 {provider === 'NAVIGRAPH' && (
                                     <div
-                                        className="p-2 mt-3 hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight rounded-md transition duration-100 cursor-pointer"
+                                        className="p-2 mt-3 rounded-md transition duration-100 cursor-pointer hover:text-theme-body bg-theme-secondary hover:bg-theme-highlight"
                                         onClick={() => dispatch(setUsingDarkTheme(!usingDarkTheme))}
                                     >
                                         {!usingDarkTheme ? <MoonFill size={40} /> : <SunFill size={40} />}
@@ -457,7 +457,7 @@ export const ChartViewer = () => {
                         </div>
 
                         <div
-                            className="flex overflow-x-hidden overflow-y-scroll relative flex-row mx-auto h-full bg-theme-accent rounded-lg grabbable no-scrollbar"
+                            className="flex overflow-x-hidden overflow-y-scroll relative flex-row mx-auto h-full rounded-lg bg-theme-accent grabbable no-scrollbar"
                             ref={ref}
                         >
                             <TransformComponent wrapperStyle={{ height: ref.current?.clientHeight, width: ref.current?.clientWidth }}>
@@ -477,7 +477,7 @@ export const ChartViewer = () => {
                                 >
                                     {(chartLinks && provider === 'NAVIGRAPH') && (
                                         <p
-                                            className="absolute top-0 left-0 font-bold text-theme-highlight whitespace-nowrap transition duration-100 transform -translate-y-full"
+                                            className="absolute top-0 left-0 font-bold whitespace-nowrap transition duration-100 transform -translate-y-full text-theme-highlight"
                                         >
                                             This chart is linked to
                                             {' '}
