+++ fbw-a32nx/src/systems/instruments/src/EFB/Navigation/Pages/NavigraphPage/NavigraphChartSelector.tsx
@@ -13,7 +13,7 @@ import {
     addPinnedChart,
 } from '../../../Store/features/navigationPage';
 import { navigationTabs } from '../../Navigation';
-import { NavigraphChart } from '../../../ChartsApi/Navigraph';
+import { NavigraphChart } from '../../../Apis/Navigraph/Navigraph';
 
 interface NavigraphChartSelectorProps {
     selectedTab: OrganizedChart;
@@ -125,7 +125,7 @@ export const NavigraphChartSelector = ({ selectedTab, loading }: NavigraphChartS
                     <>
                         {organizedCharts.map((item) => (
                             <div className="flex overflow-hidden flex-col w-full rounded-md divide-y-2 divide-gray-700" key={item.name}>
-                                <span className="p-1 text-center bg-theme-secondary rounded-t-lg">{item.name}</span>
+                                <span className="p-1 text-center rounded-t-lg bg-theme-secondary">{item.name}</span>
                                 {item.charts.map((chart) => (
                                     <div
                                         className="flex flex-row bg-theme-accent"
@@ -138,7 +138,7 @@ export const NavigraphChartSelector = ({ selectedTab, loading }: NavigraphChartS
                                                 : 'bg-theme-secondary'}`}
                                             />
                                             <div
-                                                className="flex items-center px-2 h-full hover:text-theme-body hover:bg-theme-highlight transition duration-100"
+                                                className="flex items-center px-2 h-full transition duration-100 hover:text-theme-body hover:bg-theme-highlight"
                                                 onClick={(event) => {
                                                     event.stopPropagation();
 
@@ -170,7 +170,7 @@ export const NavigraphChartSelector = ({ selectedTab, loading }: NavigraphChartS
                                         </div>
                                         <div className="flex flex-col m-2">
                                             <span>{chart.procedureIdentifier}</span>
-                                            <span className="px-2 mt-0.5 mr-auto text-sm text-theme-text bg-theme-secondary rounded-md">
+                                            <span className="px-2 mt-0.5 mr-auto text-sm rounded-md text-theme-text bg-theme-secondary">
                                                 {chart.indexNumber}
                                             </span>
                                         </div>
@@ -184,7 +184,7 @@ export const NavigraphChartSelector = ({ selectedTab, loading }: NavigraphChartS
                     <>
                         {selectedTab.charts.map((chart) => (
                             <div
-                                className="flex overflow-hidden flex-row w-full bg-theme-accent rounded-md"
+                                className="flex overflow-hidden flex-row w-full rounded-md bg-theme-accent"
                                 onClick={() => handleChartClick(chart)}
                                 key={chart.id}
                             >
@@ -194,7 +194,7 @@ export const NavigraphChartSelector = ({ selectedTab, loading }: NavigraphChartS
                                         : 'bg-theme-secondary'}`}
                                     />
                                     <div
-                                        className="flex items-center px-2 h-full hover:text-theme-body hover:bg-theme-highlight transition duration-100"
+                                        className="flex items-center px-2 h-full transition duration-100 hover:text-theme-body hover:bg-theme-highlight"
                                         onClick={(event) => {
                                             event.stopPropagation();
 
@@ -227,7 +227,7 @@ export const NavigraphChartSelector = ({ selectedTab, loading }: NavigraphChartS
                                 <div className="flex flex-col m-2">
                                     <span>{chart.procedureIdentifier}</span>
                                     <span
-                                        className="px-2 mr-auto text-sm text-theme-text bg-theme-secondary rounded-sm"
+                                        className="px-2 mr-auto text-sm rounded-sm text-theme-text bg-theme-secondary"
                                     >
                                         {chart.indexNumber}
                                     </span>
